+++ R/add_package_doc.R
@@ -6,20 +6,16 @@
 #' (where `pkg` is the name of the package). It a good place to put general 
 #' directives like `@import` and `@importFrom`.
 #'   
-#' @param open a logical value
+#' @param open A logical value. If `TRUE` (default) the file is opened in the 
+#'   editor.
 #' 
-#'   If `TRUE` (default) the file is opened in the editor.
-#' 
-#' @param overwrite a logical value
-#' 
-#'   If this file is already present and `overwrite = TRUE`, it will be erased 
-#'   and replaced. Default is `FALSE`.
+#' @param overwrite A logical value. If this file is already present and 
+#'   `overwrite = TRUE`, it will be erased and replaced. Default is `FALSE`.
 #'   
-#' @param quiet a logical value
+#' @param quiet A logical value. If `TRUE` messages are deleted. Default is 
+#'   `FALSE`.
 #' 
-#'   If `TRUE` messages are deleted. Default is `FALSE`.
-#' 
-#' @return None
+#' @return No return value.
 #'
 #' @export
 #' 
@@ -60,7 +56,7 @@ add_package_doc <- function(open = TRUE, overwrite = FALSE, quiet = FALSE) {
     dir.create(file.path(path_proj(), "R"), showWarnings = FALSE)
   
   invisible(
-    file.copy(system.file(file.path("templates", "__INDEX__"), 
+    file.copy(system.file(file.path("templates", "package-package.R"), 
                           package = "rcompendium"), path, overwrite = TRUE))