+++ R/add_to_gitignore.R
@@ -1,24 +1,20 @@
-#' Create a gitignore file
+#' Add to the .gitignore file
 #' 
 #' @description 
 #' This function creates a `.gitignore` file at the root of the project based on
 #' a template (specific to R). If a `.gitignore` is already present, files to
 #' be untracked by **git** are just added to this file.
 #' 
-#' @param x a character vector
+#' @param x A character vector. One or several files/folders names to be added 
+#'   to the `.gitignore`.
 #' 
-#'   One or several files/directories names to be added to the `.gitignore`.
+#' @param open A logical value. If `TRUE` the `.gitignore` file is opened in 
+#'   the editor. Default is `FALSE`.
 #' 
-#' @param open a logical value
+#' @param quiet A logical value. If `TRUE` messages are deleted. Default is 
+#'   `FALSE`.
 #' 
-#'   If `TRUE` the `.gitignore` file is opened in the editor.
-#'   Default is `FALSE`.
-#' 
-#' @param quiet a logical value
-#' 
-#'   If `TRUE` messages are deleted. Default is `FALSE`.
-#' 
-#' @return None
+#' @return No return value.
 #'
 #' @export
 #' 
@@ -43,7 +39,7 @@ add_to_gitignore <- function(x, open = FALSE, quiet = FALSE) {
   if (!file.exists(path)) {
     
     invisible(
-      file.copy(system.file(file.path("templates", "__GITIGNORE__"), 
+      file.copy(system.file(file.path("templates", "gitignore"), 
                             package = "rcompendium"), path))
     
     if (!quiet) ui_done("Writing {ui_value('.gitignore')} file")