+++ NEWS.md
@@ -1,3 +1,34 @@
+# rcompendium 1.0
+
+* New feature: `add_dockerfile()` creates a basic `Dockerfile` in compendium (new 
+argument `dokerfile` in `new_compendium()`) based on `rocker/rstudio`.
+* New feature: `add_renv()` initialize `renv` environment in compendium (new 
+argument `renv` in `new_compendium()`)
+* New feature: `add_github_actions_render()` will automatically render
+the `README.md` on GitHub server after each push. This action is triggered
+only if the `README.Rmd` has been modified since the last commit. Also
+add new argument `gh_render` in `new_*()` functions.
+* New vignette: developing an R package
+* New vignette: working with a compendium
+* Function `add_dependencies()` allows now missing `R/` folder and improves the
+detection of dependencies in vignettes.
+* Function `add_lifecycle_badge()` does not copy badge SVG in the project 
+anymore. The image badge is now created using https://shields.io/.
+* Update GHA templates. They are now derived from:
+https://github.com/r-lib/actions/tree/v2-branch/examples
+* Rename default vignette title (title is now _Get Started_)
+* Add `cph` (copyright holder) tag in `DESCRIPTION` file
+* Change default package hexSticker and add R script in `inst/package-sticker/` 
+to easily change the hexSticker
+* Rename commits messages (and remove emoji)
+* For compendium: arguments `gh_check` and `website` are now `FALSE` by default
+* For compendium: rename `rscripts/` folder to `analyses/`
+* For compendium: delete `paper/` folder
+* Review and improve documentation
+* Rename templates file names
+* Fix bug: detection of nested RStudio projects
+
+
 # rcompendium 0.5.1
 
 After first submission to CRAN: